requirements.txt
matplotlib
numpy
torch
neuraloperator
xarray
netCDF4
ncplot